+++ HISTORY.rst
@@ -7,6 +7,7 @@
   GitLab issue #894. Thanks to forum user @Lffy for the report.
   https://forum.mayan-edms.com/viewtopic.php?t=4027
 - Fix staging folder file get_image method.
+- Suppress staging folder file image tasks dependency errors on debug mode.
 
 3.5 (2020-09-30)
 ================